fix: adding change to router and controller -removing function unused…
and cleaning code
