+++ lib/games/game/score.ex
@@ -18,5 +18,4 @@ defmodule Games.Game.Score do
     |> validate_required([:red_card, :yellow_card, :goal, :goal_scorer])
   end
 
-
 end
