--- lib/games_web/templates/random_score/new_random.html.eex
@@ -1,5 +1,0 @@
-<h1>New Score</h1>
-
-<%= render "form_random.html", Map.put(assigns, :action, Routes.random_score_path(@conn, :create_random)) %>
-
-<span><%= link "Back", to: Routes.score_path(@conn, :index) %></span>
