--- lib/games_web/views/random_score_view.ex
@@ -1,3 +1,0 @@
-defmodule GamesWeb.RandomScoreView do
-  use GamesWeb, :view
-end
